Fixed some bug relating to a warning about labeling of training data and not doing the same to the test data.
Used textblob to classify ratings instead of the previous assumption
approach. Found that there were about 500 discrepancies in our ratings.

Build a classifier model using `MultinomialNB()`. The accuracy was lower
but still predicted for all the ratings categories.

When the model was tested using a negative comment, the model predicted
it right.